fix #47 :extract sources file and fix data race on ff variable
